dsfsd

+++ app/Http/Controllers/CustomerController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Call;
 use App\Customer;
+use App\Plan;
 use App\Task;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
@@ -43,8 +44,15 @@ class CustomerController extends Controller
      */
     public function store(Request $request)
     {
+        if($request->id){
         $call = Call::find($request->id);
         $call->delete();
+            $today = Carbon::now()->setTime('00', '00');
+            $plan = Plan::where('created_at', '>=', $today)->where('user_id',auth()->id())->where('status',null)->first();
+            $plan->calls_score = $plan->calls_score + 1;
+            $plan->save();
+        }
+
         $customer = New Customer();
         $customer->name = $request->name;
         $customer->company = $request->company;
@@ -54,39 +62,34 @@ class CustomerController extends Controller
 
         $task = new Task();
         $task->title = $customer->name ? $customer->company : 'Empty';
-        $request->request->remove('date');
         $deadline_date = Carbon::parseFromLocale($request->date, 'ru');
         $request->request->remove('date');
         $request->merge(['date' => $deadline_date]);
         $task->deadline_date = $request->date;
         $task->user_id = auth()->check() ? auth()->id() : 0;
 
-        if($request->status == "true")
+        $task->save();
+        $customer->task()->save($task);
+        if($request->id)
         {
-            $task->status_id = 1;
+            if ($request->ajax()){
+                return response()->json([
+                    'status' => "success",
+                    'data' => $task,
+                    'plan' => $plan
+                ], 200);
+            }
         }
         else
         {
-            $task->status_id = 0;
+            if ($request->ajax()){
+                return response()->json([
+                    'status' => "success",
+                    'data' => $task
+                ], 200);
+            }
         }
 
-        $task->save();
-        $customer->task()->save($task);
-        if ($request->ajax() && $task->status_id == 1){
-            return response()->json([
-                'status' => "success",
-                'data' => $task,
-                'view' => view('tasks.potentials-card', [
-                    'customer' => $task,
-                ])->render(),
-            ], 200);
-        }
-        else {
-            return response()->json([
-                'status' => "success",
-                'data' => $task
-            ]);
-        }
 
         return back();
     }
@@ -120,9 +123,32 @@ class CustomerController extends Controller
      * @param  \App\Customer  $customer
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Customer $customer)
+    public function update(Request $request)
     {
-        //
+        $customer = Customer::find($request->id);
+        $task = $customer->task;
+        $task->title = $request->name;
+        $customer->name = $request->name;
+        $customer->company = $request->company;
+        $customer->contacts = $request->phone;
+        $customer->save();
+        $task->save();
+        if(isset($customer->meeting->id)) {
+            $id = $customer->meeting->id;
+        }
+        else
+        {
+            $id = 0;
+        }
+        if ($request->ajax()){
+            return response()->json([
+                'status' => "success",
+                'data' => $customer,
+                'id' => $id,
+            ]);
+        }
+
+        return back();
     }
 
     /**
@@ -133,10 +159,10 @@ class CustomerController extends Controller
      */
     public function delete(Request $request)
     {
-        $task = Task::find($request->id);
-        $customer = $task->taskable;
-        $customer->delete();
-        $customer->delete();
+        $customer = Customer::find($request->id);
+        $task = $customer->task;
+        $task->status_id = 0;
+        $task->save();
 
         if ($request->ajax()){
             return response()->json([
@@ -150,5 +176,30 @@ class CustomerController extends Controller
     public function destroy(Customer $customer)
     {
         //
+    }
+
+    public function change(Request $request)
+    {
+        $customer = Customer::find($request->id);
+        $task = $customer->task;
+        $task->description = $request->desc;
+        $deadline_date = Carbon::parseFromLocale($request->date, 'ru');
+        $request->request->remove('date');
+        $request->merge(['date' => $deadline_date]);
+        $task->deadline_date = $request->date;
+        $task->status_id = 1;
+        $task->save();
+
+        if ($request->ajax()){
+            return response()->json([
+                'status' => "success",
+                'data' => $task,
+                'view' => view('tasks.potentials-card', [
+                    'customer' => $task,
+                ])->render(),
+            ], 200);
+        }
+
+        return back();
     }
 }
